Hide voice search button if intent cannot be handled
If the Google Search app is disabled (this only seems to be possible
on Nexus 4), the Dialer crashes if the voice search button is pressed.
This CL hides/shows the voice search button every time the dialer
is launched/resumed after checking to see if there exists an activity
that can handle the ACTION_RECOGNIZE_SPEECH intent.

A string to indicate that voice search is unavailable is also added
just in case the user ends up in a weird state where the button is
showing but clicking on it would throw an ActivityNotFoundException.

Bug: 12015318
Change-Id: Idd7ec2da422425dd95ae0060ebc9b85a2cf35fb0
(cherry picked from commit da0f9046b2457c8adc0b97896d7ea1e880eccf96)

Conflicts:
	src/com/android/dialer/DialtactsActivity.java